requirements.txt
certifi==2021.10.8
charset-normalizer==2.0.9
guppy3==3.1.2
idna==3.3
memory-profiler==0.59.0
psutil==5.8.0
requests==2.26.0
urllib3==1.26.7
numpy~=1.21.4